# feat: cache for prices